// src/types/setting-Types.ts
export interface UserPermissionDetails {
  userId: string | User
  allowedFields: string[]
  read: boolean
  write: boolean
  update: boolean
  delete: boolean
  _id?: string // Optional, as it might be generated by the backend
}

export interface AccessPermissionRole {
  role: string
  permissions: UserPermissionDetails[]
  _id?: string // Optional, as it might be generated by the backend
}

export interface PermissionModule {
  _id: string
  module: string
  AccessPermissions: AccessPermissionRole[]
  createdAt: string
  updatedAt: string
  __v: number
  updatedBy?: string | User | null
}

// Add a User interface to handle populated user data
export interface User {
  _id: string
  email: string
  phone_Number: string
  role: string
  address: any[] // Define a more specific type if needed
  __v: number
  ticketsAssigned?: any[] // Optional
  cartId?: string | null // Optional
  fcmToken?: string | null // Optional
  wishlistId?: string | null // Optional
  last_login?: string // Optional
  vehicle_details?: any[] // Optional
}